payment_project/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-payments==3.0.1
django-phonenumber-field==8.1.0
idna==3.10
phonenumberslite==9.0.10
psycopg2-binary==2.9.10
requests==2.32.4
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
